+++ OpenBCI_32bit_Library.cpp
@@ -597,7 +597,7 @@ boolean OpenBCI_32bit_Library::boardBeginDebug(uint8_t baudRate) {
 */
 void OpenBCI_32bit_Library::boardReset(void) {
   initialize(); // initalizes accelerometer and on-board ADS and on-daisy ADS if present
-  // delay(500);
+  delay(500);
   configureLeadOffDetection(LOFF_MAG_6NA, LOFF_FREQ_31p2HZ);
   Serial0.println("OpenBCI V3 8-16 channel");
   Serial0.print("On Board ADS1299 Device ID: 0x"); Serial0.println(ADS_getDeviceID(ON_BOARD),HEX);
@@ -607,6 +607,8 @@ void OpenBCI_32bit_Library::boardReset(void) {
   Serial0.print("LIS3DH Device ID: 0x"); Serial0.println(LIS3DH_getDeviceID(),HEX);
   Serial0.println("Firmware: v3.0.0");
   sendEOT();
+  delay(5);
+  wifiReset();
 }
 
 /**
@@ -635,6 +637,10 @@ void OpenBCI_32bit_Library::activateAllChannelsToTestCondition(byte testInputCod
   // Restart stream if need be
   if (wasStreaming) {
     streamStart();
+  } else {
+    printSuccess();
+    printAll("Configured internal test signal.");
+    sendEOT();
   }
 }
 
@@ -963,9 +969,8 @@ void OpenBCI_32bit_Library::initialize(){
   pinMode(BOARD_ADS, OUTPUT); digitalWrite(BOARD_ADS,HIGH);
   pinMode(DAISY_ADS, OUTPUT); digitalWrite(DAISY_ADS,HIGH);
   pinMode(LIS3DH_SS,OUTPUT); digitalWrite(LIS3DH_SS,HIGH);
-  pinMode(WIFI_SS,OUTPUT); digitalWrite(WIFI_SS, LOW);
+  pinMode(WIFI_SS, INPUT);
   pinMode(WIFI_RESET,OUTPUT); digitalWrite(WIFI_RESET, LOW);
-  wifiReset();
 
   spi.begin();
   spi.setSpeed(4000000);  // use 4MHz for ADS and LIS3DH
@@ -980,36 +985,42 @@ void OpenBCI_32bit_Library::initialize(){
  */
 void OpenBCI_32bit_Library::loop(void) {
   if (toggleWifiReset) {
-    if ((millis() - timeOfWifiToggle) > 100) {
-      digitalWrite(WIFI_SS, LOW);
-      delay(10);
+    if (millis() > timeOfWifiToggle + 500) {
       digitalWrite(WIFI_RESET, HIGH);
       toggleWifiReset = false;
+      toggleWifiCS = true;
     }
   }
   if (toggleWifiCS) {
-    if ((millis() - timeOfWifiToggle) > 2000) {
+    if (millis() > timeOfWifiToggle + 3000) {
       digitalWrite(OPENBCI_PIN_LED, HIGH);
-      digitalWrite(WIFI_SS, HIGH); // Set back to high
+      pinMode(WIFI_SS, OUTPUT); digitalWrite(WIFI_SS, HIGH); // Set back to high
       toggleWifiCS = false;
       timeOfWifiToggle = millis();
       seekingWifi = true;
     }
   }
   if (seekingWifi) {
-    if ((millis() - timeOfWifiToggle) > 3000) {
+    if (millis() > timeOfWifiToggle + 4000) {
       seekingWifi = false;
+      Serial0.print("seeking wifi");
       if (!wifiAttach()) {
+        Serial0.println(" not attached");
         wifiAttachAttempts++;
-        if (wifiAttachAttempts < 3) {
+        if (wifiAttachAttempts < 10) {
           seekingWifi = true;
           timeOfWifiToggle = millis();
         }
+      } else {
+        Serial0.println(" attached sending gains");
+
+        // Send gains
+        wifiSendGains();
       }
     }
   }
   if (iWifi.rx) {
-    if ((millis() - timeOfLastRead) > 20) {
+    if (millis() > timeOfLastRead + 20) {
       wifiReadData();
       uint8_t numChars = (uint8_t)wifiBufferRx[0];
       if (numChars > 0 && numChars < WIFI_SPI_MAX_PACKET_SIZE) {
@@ -3134,13 +3145,11 @@ boolean OpenBCI_32bit_Library::wifiRemove(void) {
 void OpenBCI_32bit_Library::wifiReset(void) {
   // Always keep pin low or else esp will fail to boot.
   // See https://github.com/esp8266/Arduino/blob/master/libraries/SPISlave/examples/SPISlave_SafeMaster/SPISlave_SafeMaster.ino#L12-L15
-  digitalWrite(WIFI_SS,LOW);
-  delay(1);
+  pinMode(WIFI_SS, INPUT);
   digitalWrite(WIFI_RESET, LOW); // Reset the ESP8266
   digitalWrite(OPENBCI_PIN_LED, LOW); // Good visual indicator of what's going on
   initializeSpiInfo(iWifi);
   wifiPresent = false;
-  toggleWifiCS = true;
   toggleWifiReset = true;
   timeOfWifiToggle = millis();
 }
@@ -3153,6 +3162,7 @@ void OpenBCI_32bit_Library::wifiSendGains(void) {
   wifiBufferTxClear();
 
   wifiStoreByteBufTx(WIFI_SPI_MSG_GAINS);
+  wifiStoreByteBufTx(WIFI_SPI_MSG_GAINS); // Redundancy
   wifiStoreByteBufTx(numChannels);
   for (uint8_t i = 0; i < numChannels; i++) {
     wifiStoreByteBufTx(channelSettings[i][GAIN_SET]);
@@ -3223,12 +3233,13 @@ void OpenBCI_32bit_Library::wifiSendStringMulti(const char *str) {
 boolean OpenBCI_32bit_Library::wifiSmell(void){
   boolean isWifi = false;
   uint32_t uuid = wifiReadStatus();
-  if (verbosity) {
+  // if (verbosity) {
     Serial0.print("Wifi ID 0x");
     Serial0.println(uuid,HEX);
     Serial.println(micros());
     sendEOT();
-  }
+  // }
+
   if(uuid == 209) {isWifi = true;} // should read as 0x3E
   return isWifi;
 }